pyproject.toml
[project]
name = "mdtyp"
version = "0.1.0"
description = "Convert Markdown documents to Typst"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "markdown-it-py>=4.0.0",
    "mdit-py-plugins>=0.5.0",
    "typer>=0.24.1",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
mdtyp = "mdtyp.cli:app"
